Redesign and refactor regular tuple template
This commit removes the old design of a regular tuple which uses
multiple inheritance and wrapper classes for the given types.
Due to alignment constraints of base classes, this design took much
more storage than expected even for simple tuple instantiations.
Furthermore, the possible property of virtual inheritance of given types
could introduce a lot more problems.
The benefit of EBCO to remove the size increase of empty classes
is not enough to weigh up the disadvantages of the above problems.